Reorganized a bit. Also, should be my actual account instead of an old one I forgot about.
